+++ src/Ydb.Sdk/src/Tracing/YdbActivitySource.cs
@@ -16,9 +16,20 @@ internal static class YdbActivitySource
         if (exception is YdbException ydbException)
         {
             activity.SetTag("db.response.status_code", ydbException.Code);
+            activity.SetTag("error.type", ydbException.Code
+                is StatusCode.ClientTransportUnknown
+                or StatusCode.ClientTransportUnavailable
+                or StatusCode.ClientTransportTimeout
+                or StatusCode.ClientTransportResourceExhausted
+                or StatusCode.ClientTransportUnimplemented
+                ? "transport_error"
+                : "ydb_error");
+        }
+        else
+        {
+            activity.SetTag("error.type", exception.GetType().FullName);
         }
 
-        activity.SetTag("error.type", exception.GetType().FullName);
         activity.SetStatus(ActivityStatusCode.Error, exception.Message);
     }